modificacion del motor para operaciones de insercion
